Fix modal Select automatically closing
