+++ packages/ariakit/src/dialog/__utils/use-previous-mouse-down-ref.ts
@@ -5,7 +5,10 @@ export function usePreviousMouseDownRef(enabled?: boolean) {
   const previousMouseDownRef = useRef<EventTarget | null>();
 
   useEffect(() => {
-    if (!enabled) return;
+    if (!enabled) {
+      previousMouseDownRef.current = null;
+      return;
+    }
     const onMouseDown = (event: MouseEvent) => {
       previousMouseDownRef.current = event.target;
     };
